chore(suite-native): rename trading-common to trading-quote-utils
