+++ jest.config.native.js
@@ -55,7 +55,7 @@ module.exports = {
         '<rootDir>/../../suite-native/react-native-graph/src/jestSetup.js',
         '<rootDir>/../../suite-native/atoms/src/jestSetup.jsx',
         '<rootDir>/../../suite-native/module-trading/src/jest.setup.tsx',
-        '<rootDir>/../../suite-native/trading-common/src/jest.setup.tsx',
+        '<rootDir>/../../suite-native/trading-quote-utils/src/jest.setup.tsx',
         '<rootDir>/../../suite-native/module-connect-popup/src/jest.setup.ts',
         '<rootDir>/../../suite-native/module-device-onboarding/src/jest.setup.ts',
         '<rootDir>/../../suite-native/config/src/jest.setup.ts',
