+++ suite-native/module-trading/package.json
@@ -66,11 +66,11 @@
         "@suite-native/trading-analytics": "workspace:*",
         "@suite-native/trading-atoms": "workspace:*",
         "@suite-native/trading-browser-auth": "workspace:*",
-        "@suite-native/trading-common": "workspace:*",
         "@suite-native/trading-consts": "workspace:*",
         "@suite-native/trading-debug": "workspace:*",
         "@suite-native/trading-history": "workspace:*",
         "@suite-native/trading-provider-utils": "workspace:*",
+        "@suite-native/trading-quote-utils": "workspace:*",
         "@suite-native/trading-residence": "workspace:*",
         "@suite-native/trading-state": "workspace:*",
         "@suite-native/transaction-management": "workspace:*",
