+++ suite-native/module-trading/src/components/exchange/ExchangePreview/ExchangeFromAccountTradePreviewCard.tsx
@@ -5,7 +5,10 @@ import type { ExchangeTrade } from 'invity-api';
 import { Text } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
 import { TradeSideCard } from '@suite-native/trading-atoms';
-import { CryptoToFiatValueBadge, useChangeStringsExtractor } from '@suite-native/trading-common';
+import {
+    CryptoToFiatValueBadge,
+    useChangeStringsExtractor,
+} from '@suite-native/trading-quote-utils';
 import { selectExchangeSelectedSendAccount } from '@suite-native/trading-state';
 
 export type ExchangeFromAccountTradePreviewCardProps = {
