+++ suite-native/module-trading/src/components/sell/SellPreview/SellToFiatTradePreviewCard.tsx
@@ -2,7 +2,7 @@ import type { FiatCurrencyCode, SellFiatTrade } from 'invity-api';
 
 import { Text } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
-import { useChangeStringsExtractor } from '@suite-native/trading-common';
+import { useChangeStringsExtractor } from '@suite-native/trading-quote-utils';
 
 import { TradeFiatSideCard } from '../../general/TradeInfo/TradeFiatSideCard';
