+++ suite-native/module-trading/src/hooks/buy/__tests__/useBuyData.test.tsx
@@ -16,8 +16,8 @@ import { tradingSlice } from '@suite-native/trading-state';
 
 import { useBuyData } from '../useBuyData';
 
-jest.mock('@suite-native/trading-common', () => ({
-    ...jest.requireActual('@suite-native/trading-common'),
+jest.mock('@suite-native/trading-quote-utils', () => ({
+    ...jest.requireActual('@suite-native/trading-quote-utils'),
     getRandomAccountDescriptor: () => 'random_string',
 }));
