+++ suite-native/module-trading/src/hooks/buy/useBuyData.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { selectTradingBuyLoadingTimestampAndStatus, tradingThunks } from '@suite-common/trading';
-import { getRandomAccountDescriptor } from '@suite-native/trading-common';
+import { getRandomAccountDescriptor } from '@suite-native/trading-quote-utils';
 import { selectBuySelectedReceiveAccount } from '@suite-native/trading-state';
 
 export const useBuyData = (reloadRequestOrdinal: number) => {
