+++ suite-native/module-trading/src/hooks/sell/__tests__/useSellData.test.tsx
@@ -6,8 +6,8 @@ import { getBtcAccount, getInitializedTradingState } from '@suite-native/trading
 import { createTradingLightStore } from '../../../__tests__/tradingTestUtils';
 import { useSellData } from '../useSellData';
 
-jest.mock('@suite-native/trading-common', () => ({
-    ...jest.requireActual('@suite-native/trading-common'),
+jest.mock('@suite-native/trading-quote-utils', () => ({
+    ...jest.requireActual('@suite-native/trading-quote-utils'),
     getRandomAccountDescriptor: () => 'random_string',
 }));
