+++ suite-native/module-trading/src/hooks/sell/useSellData.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { selectTradingSellLoadingTimestampAndStatus, tradingThunks } from '@suite-common/trading';
-import { getRandomAccountDescriptor } from '@suite-native/trading-common';
+import { getRandomAccountDescriptor } from '@suite-native/trading-quote-utils';
 import { selectSellSelectedSendAccount } from '@suite-native/trading-state';
 
 export const useSellData = (reloadRequestOrdinal: number) => {
