+++ suite-native/module-trading/tsconfig.json
@@ -75,11 +75,11 @@
         { "path": "../trading-analytics" },
         { "path": "../trading-atoms" },
         { "path": "../trading-browser-auth" },
-        { "path": "../trading-common" },
         { "path": "../trading-consts" },
         { "path": "../trading-debug" },
         { "path": "../trading-history" },
         { "path": "../trading-provider-utils" },
+        { "path": "../trading-quote-utils" },
         { "path": "../trading-residence" },
         { "path": "../trading-state" },
         { "path": "../transaction-management" },
