+++ suite-native/trading-history/package.json
@@ -29,8 +29,8 @@
         "@suite-native/link": "workspace:*",
         "@suite-native/trading-atoms": "workspace:*",
         "@suite-native/trading-browser-auth": "workspace:*",
-        "@suite-native/trading-common": "workspace:*",
         "@suite-native/trading-provider-utils": "workspace:*",
+        "@suite-native/trading-quote-utils": "workspace:*",
         "@suite-native/trading-state": "workspace:*",
         "@trezor/styles-native": "workspace:*",
         "@trezor/type-utils": "workspace:*",
