+++ suite-native/trading-history/src/components/TradeDetailSheet/TradeDetailAlert.tsx
@@ -11,7 +11,7 @@ import { FullAlertBox } from '@suite-native/atoms';
 import type { IconName } from '@suite-native/icons';
 import { type TxKeyPath, useTranslate } from '@suite-native/intl';
 import { buildTradingUrl, useBrowserAuth } from '@suite-native/trading-browser-auth';
-import type { TradeStatusStep } from '@suite-native/trading-common';
+import type { TradeStatusStep } from '@suite-native/trading-quote-utils';
 import { exhaustive } from '@trezor/type-utils';
 
 type AlertConfig = {
