+++ suite-native/trading-history/src/components/TradeDetailSheet/TradeDetailSheet.tsx
@@ -4,8 +4,8 @@ import { useSelector } from 'react-redux';
 import { type TradingRootState, selectTradingTradeByOrderId } from '@suite-common/trading';
 import { BottomSheetModal, useBottomSheetModal } from '@suite-native/atoms';
 import { useTranslate } from '@suite-native/intl';
-import { getTradeTitle } from '@suite-native/trading-common';
 import { Footer } from '@suite-native/trading-provider-utils';
+import { getTradeTitle } from '@suite-native/trading-quote-utils';
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
 
 import { TradeDetailFooter } from './TradeDetailFooter';
