+++ suite-native/trading-history/src/components/TradeDetailSheet/__tests__/TradeDetailAlert.test.tsx
@@ -1,8 +1,8 @@
 import { type TradingTransaction } from '@suite-common/trading';
 import { type TxKeyPath, getTranslation } from '@suite-native/intl';
 import { renderWithStoreProvider } from '@suite-native/test-utils-store';
-import { type TradeStatusStep } from '@suite-native/trading-common';
 import { getBuyTrade, getInitializedTradingState } from '@suite-native/trading-fixtures';
+import { type TradeStatusStep } from '@suite-native/trading-quote-utils';
 
 import { TradeDetailAlert } from '../TradeDetailAlert';
