+++ suite-native/trading-quote-utils/src/components/CryptoToFiatValueBadge.tsx
