+++ suite-native/trading-quote-utils/src/components/__tests__/CryptoToFiatValueBadge.test.tsx
