+++ suite-native/trading-quote-utils/src/hooks/__tests__/useChangeStringsExtractor.test.ts
