+++ suite-native/trading-quote-utils/src/hooks/useChangeStringsExtractor.ts
