+++ suite-native/trading-quote-utils/src/hooks/useTradingFiatValues.ts
