+++ suite-native/trading-quote-utils/src/utils/utils.ts
