+++ suite-native/trading-quote-utils/tsconfig.json
@@ -17,7 +17,6 @@
         },
         { "path": "../atoms" },
         { "path": "../intl" },
-        { "path": "../trading-types" },
         { "path": "../../packages/type-utils" },
         { "path": "../../packages/utils" },
         { "path": "../test-utils" },
